+++ NEWS.md
@@ -1,6 +1,14 @@
+# dreamlet 0.0.5
+	August 25, 2021
+	- handling of pmetadata by processAssays(), fitVarPart(), and dreamlet()
+	- change defaults for bpparam to SerialParam()
+	- Created new files for code
+	- Create object dreamletResult returned by dreamlet() and used by topTable()
+	- more capable dreamletProcessedData object
+
+
 # dreamlet 0.0.4
 	- add aggregateToPseudoBulk() for faster access to SingleCellExperiment() backed by H5AD
-
 # dreamlet 0.0.3
 	- add argument to `processAssays()` to include extra meta-data	
 	- add subseting with assay() for dreamletProcessedData